+++ final/Foundation4/Running.cs
@@ -1,10 +1,10 @@
 class Running : Activity
 {
 
-private int _distance;
+private double _distance;
 
 
-public Running(string date, int length, int distance) : base(date, length)
+public Running(string date, int length, double distance) : base(date, length)
     {
         _distance = distance;
     }
